+++ tap_sybase/sync_strategies/full_table.py
@@ -55,12 +55,12 @@ def sync_table(mssql_conn, config, catalog_entry, state, columns, stream_version
 
     with connect_with_backoff(mssql_conn) as open_conn:
         with open_conn.cursor() as cur:
-            select_sql = common.generate_select_sql(catalog_entry, columns)
+            select_sql = common.generate_select_sql(catalog_entry, columns, config)
 
             params = {}
 
             common.sync_query(
-                cur, catalog_entry, state, select_sql, columns, stream_version, params, config
+                cur, catalog_entry, state, select_sql, columns, stream_version, params
             )
 
     # clear max pk value and last pk fetched upon successful sync